backend/requirements.txt
numpy
requests
Flask
Werkzeug
faiss-cpu
redis
tqdm
pymongo
flask-cors
pyOpenSSL
Flask-SocketIO
Pillow
scikit-learn
loguru
matplotlib